+++ main/src/components/ListResultTable/TemplatedListResultTable.jsx
@@ -21,8 +21,9 @@ const TemplatedListResultTable = (props) => {
   const location = useLocation();
   const navigate = useNavigate();
   const query = configManager.getQueryWorkingCopyById(resource);
-  const [waitingForVariableOptions, setWaitingForVariableOptions] = useState(!!(query.variables || query.indirectVariables));
-  const [waitingForVariableOptionsError, setWaitingForVariableOptionsError] = useState("");
+  const [askingForVariableOptions, setAskingForVariableOptions] = useState(false);
+  const [waitingForVariableOptions, setWaitingForVariableOptions] = useState(false);
+  const [variableOptionsError, setVariableOptionsError] = useState("");
   const [variableOptions, setVariableOptions] = useState({});
   const [variableValues, setVariableValues] = useState({});
   const [variablesSubmitted, setVariablesSubmitted] = useState(false);
@@ -32,8 +33,9 @@ const TemplatedListResultTable = (props) => {
   // LOG console.log(`--- TemplatedListResultTable #${++templatedListResultTableCounter}`);
   // LOG console.log(`props: ${JSON.stringify(props, null, 2)}`);
   // LOG console.log(`resource: ${resource}`);
+  // LOG console.log(`askingForVariableOptions: ${askingForVariableOptions}`);
   // LOG console.log(`waitingForVariableOptions: ${waitingForVariableOptions}`);
-  // LOG console.log(`waitingForVariableOptionsError: ${waitingForVariableOptionsError}`);
+  // LOG console.log(`variableOptionsError: ${variableOptionsError}`);
   // LOG console.log(`variableOptions: ${JSON.stringify(variableOptions, null, 2)}`);
   // LOG console.log(`variableValues: ${JSON.stringify(variableValues, null, 2)}`);
   // LOG console.log(`variablesSubmitted: ${variablesSubmitted}`);
@@ -42,21 +44,23 @@ const TemplatedListResultTable = (props) => {
 
   useEffect(() => {
     (async () => {
-      if (query.variables || query.indirectVariables) {
+      if (askingForVariableOptions) {
+        setAskingForVariableOptions(false);
         try {
-          // LOG console.log('start waiting for variable options');
+          // LOG console.log('Start waiting for variable options');
+          setWaitingForVariableOptions(true);
           // LOG const t1 = Date.now();
           setVariableOptions(await dataProvider.getVariableOptions(query));
           // LOG const t2 = Date.now();
-          // LOG console.log(`done waiting for variable options after ${t2-t1} ms`);
+          // LOG console.log(`Done waiting for variable options after ${t2-t1} ms`);
           setWaitingForVariableOptions(false);
         } catch (error) {
-          // LOG console.log(`error getting variable options: ${error.message}`);
-          setWaitingForVariableOptionsError(error.message);
+          // LOG console.log(`Error getting variable options: ${error.message}`);
+          setVariableOptionsError(error.message);
         }
       }
     })();
-  }, [resource]);
+  }, [askingForVariableOptions]);
 
 
   // Cover a transient state after creation of a new custom query. EventEmitter's event processing may still be in progress.
@@ -65,29 +69,41 @@ const TemplatedListResultTable = (props) => {
     return false;
   }
 
-  if (waitingForVariableOptionsError) {
-    // LOG console.log(`TemplatedListResultTable failure while waiting for variable options: ${waitingForVariableOptionsError}`);
-    return <ErrorDisplay errorMessage={waitingForVariableOptionsError} />;
+  if (variableOptionsError) {
+    // LOG console.log(`TemplatedListResultTable variable options error: ${variableOptionsError}`);
+    return <ErrorDisplay errorMessage={variableOptionsError} />;
+  }
+
+  if (isTemplatedQuery) {
+    if (!Object.keys(variableOptions).length) {
+      // Check for initial visit with url search parameters
+      if (!askingForVariableOptions && !waitingForVariableOptions && !variablesSubmitted) {
+        const vv = initialVariableValuesFromUrlSearchParams(new URLSearchParams(location.search));
+        if (Object.keys(vv).length) {
+          // LOG console.log("Accepting initial variable values from url search parameters.");
+          setVariableValues(vv);
+          setVariablesSubmitted(true);
+          return false;
+        }
+        // LOG console.log("Trigger the search for variable options in body.");
+        setAskingForVariableOptions(true);
+        return false;
+      }
+    } else {
+      // Check for next visit with url search parameters
+      const vv = newVariableValuesFromUrlSearchParams(new URLSearchParams(location.search), variableValues);
+      if (vv && Object.keys(variableValues).some((v) => variableValues[v] != vv[v])) {
+        // LOG console.log("Accepting new variable values from url search parameters.");
+        setVariableValues(vv);
+        setVariablesSubmitted(true);
+        return false;
+      }
+    }
   }
 
   if (waitingForVariableOptions) {
     // LOG console.log('TemplatedListResultTable waiting for variable options.');
     return <Loading sx={{ height: "auto" }} loadingSecondary={"The options for the variables in this query are loading. Just a moment please."} />;
-  }
-
-  if (isTemplatedQuery) {
-    // Check if an update of variable values is needed from user supplied url search parameters
-    const possibleNewVariableValues = variableValuesFromUrlSearchParams(new URLSearchParams(location.search), variableOptions);
-    // Protect against incomplete or omitted variable values, as is the case when changing pagination,
-    // where List causes a revisit but does not include variable values in url search parameters
-    if (Object.keys(possibleNewVariableValues).length == Object.keys(variableOptions).length) {
-      if (Object.keys(variableOptions).some((v) => variableValues[v] != possibleNewVariableValues[v])) {
-        // LOG console.log("Accepting new variable values from user supplied url search parameters.");
-        setVariableValues(possibleNewVariableValues);
-        setVariablesSubmitted(true);
-        return false;
-      }
-    }
   }
 
   const submitVariables = (formVariables) => {
@@ -102,6 +118,12 @@ const TemplatedListResultTable = (props) => {
 
   const changeVariables = () => {
     setVariablesSubmitted(false);
+    if (!Object.keys(variableOptions).length) {
+      if (!askingForVariableOptions) {
+        // LOG console.log("Trigger the search for variable options in changeVariables.");
+        setAskingForVariableOptions(true);
+      }
+    }
     // revisit with same search parameters
     navigate(location.search);
   }
@@ -129,19 +151,35 @@ function urlSearchParamsFromVariableValues(variableValues) {
 }
 
 /**
- * Make variableValues from urlSearchParams
+ * Make initial variableValues from urlSearchParams
  * @param {URLSearchParams} urlSearchParams 
- * @param {Object} variableOptions used to filter
  * @returns {Object} variableValues
  */
-function variableValuesFromUrlSearchParams(urlSearchParams, variableOptions) {
+function initialVariableValuesFromUrlSearchParams(urlSearchParams) {
   const variableValues = {};
-  for (const variableName of Object.keys(variableOptions)) {
-    if (urlSearchParams.has(variableName)) {
-      variableValues[variableName] = urlSearchParams.get(variableName);
-    }
+  for (const [key, value] of urlSearchParams.entries()) {
+    variableValues[key] = value;
   }
   return variableValues;
+}
+
+/**
+ * Make new variableValues from urlSearchParams
+ * @param {URLSearchParams} urlSearchParams the input
+ * @param {Object} variableValues the current variableValues
+ * @returns {Object} new variableValues or undefined, if not all variable keys in the input
+ */
+function newVariableValuesFromUrlSearchParams(urlSearchParams, variableValues) {
+  const newVariableValues = {};
+  for (const variableName of Object.keys(variableValues)) {
+    if (urlSearchParams.has(variableName)) {
+      newVariableValues[variableName] = urlSearchParams.get(variableName);
+    }
+  }
+  if (Object.keys(newVariableValues).length == Object.keys(variableValues).length) {
+    return newVariableValues;
+  }
+  return undefined;
 }
 
 export default TemplatedListResultTable;